break out the fuzzer logic and the script logic
